feat: добавить контекст. Вынести авторизацию и регистрацию в App
