+++ src/contexts/CurrentUserContext.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+// Создаём объект контекста и экспортируем его
+
+// eslint-disable-next-line import/prefer-default-export
+export const CurrentUserContext = React.createContext();
